Initial code checkin.
This is just a skeleton of the code. I haven't 
really thought about the design too much. But
this shows the line of thinking about how it
can be abstracted to allow development on the 
desktop without the need for a Android environment.
This should allow for anyone that knows Java to
help out. It should also open up to using this
library for java client or server.